feat(ui): better folding
