+++ lua/utils/ui.lua
@@ -47,4 +47,21 @@ M.diagnostics   = {
 }
 -- stylua: ignore end
 
+M.get_statuscolumn = function()
+  local fc = vim.opt.fillchars:get()
+  local function get_fold(lnum)
+    if vim.fn.foldlevel(lnum) <= vim.fn.foldlevel(lnum - 1) then
+      return " "
+    end
+    return vim.fn.foldclosed(lnum) == -1 and fc.foldopen or fc.foldclose
+  end
+
+  if vim.opt_local.signcolumn:get() == "yes" then
+    -- fold info + line number
+    return "%s%l%= " .. get_fold(vim.v.lnum) .. " "
+  else
+    return ""
+  end
+end
+
 return M
